requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0.5
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.0.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyotp==2.6.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.0.1
tzdata==2022.7
urllib3==1.26.14
